+++ CHANGES.md
@@ -2,6 +2,7 @@
 - Q.js lib for promises
 - almond.js lib for AMD shim
 - underscore.js for internal utility
+- F2.Interfaces for class validation
 
 # Removed
 - jQuery
@@ -15,5 +16,6 @@
 
 # Modified
 - F2.guid: now RFC4122 compliant and doesn't vary length by browser
-- F2.registerApps: now called `load`.
+- F2.registerApps: renamed to `load`.
 - F2.load: removed ability to pass in predefined app manifests
+- F2.init: renamed to `F2.config`.